docker_images/octave/app/requirements.txt
flask
scipy
google-cloud-storage